+++ src/sasctl/utils/astore.py
@@ -350,10 +350,13 @@ def _get_model_properties(result):
     else:
         algorithm = None
 
+    def is_classification(r):
+        return classification_target(r) is not None
+
     def classification_target(r):
-        target = r.OutputVariables.Name.str.startswith('I_')
-        target = r.OutputVariables.Name[target].iloc[0]
-        return target.replace('I_', '', 1)
+        target = r.OutputVariables.Name[r.OutputVariables.Name.str.startswith('I_')]
+        if target.shape[0] > 0:
+            return target.iloc[0].replace('I_', '', 1)
 
     def regression_target(r):
         target = r.OutputVariables.Name.str.startswith('P_')
@@ -375,7 +378,7 @@ def _get_model_properties(result):
     elif algorithm == 'forest':
         properties['algorithm'] = 'Random forest'
 
-        if 'Classification' in result.InputVariables.Type.values:
+        if is_classification(result):
             properties['function'] = 'classification'
             properties['targetVariable'] = classification_target(result)
         else:
@@ -385,7 +388,7 @@ def _get_model_properties(result):
     elif algorithm == 'gradboost':
         properties['algorithm'] = 'Gradient boosting'
 
-        if 'Classification' in result.InputVariables.Type.values:
+        if is_classification(result):
             properties['function'] = 'classification'
             properties['targetVariable'] = classification_target(result)
 
@@ -398,7 +401,7 @@ def _get_model_properties(result):
     elif algorithm == 'svmachine':
         properties['algorithm'] = 'Support vector machine'
 
-        if 'Classification' in result.InputVariables.Type.values:
+        if is_classification(result):
             properties['function'] = 'classification'
             properties['targetVariable'] = classification_target(result)
             properties['targetLevel'] = 'binary'