+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
- -
+**Bugfixes**
+ - Fixed an issue with `register_model()` where random forest, gradient boosting, and SVM regression models with 
+ nominal inputs where incorrectly treated as classification models. 
  
 v1.6.1 (2021-09-01)
 -------------------